--- Server-master/README.md
@@ -1,2 +1,0 @@
-# Server
-服务端程序
