--- Server-master/main.cpp
@@ -1,11 +1,0 @@
-#include "server.h"
-#include <QApplication>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-    Server w;
-    w.show();
-
-    return a.exec();
-}
