+++ attribute.h
@@ -3,6 +3,9 @@
 
 #include <QJsonObject>
 #include <QString>
+#include <QJsonDocument>
+
+#include <QtNetwork>
 
 class Attribute
 {
@@ -14,39 +17,41 @@ public:
     //static int MODE_HEAT = 1;           //工作模式制热
 
 public:
-    Attribute();                        //构造函数
-    ~Attribute();                       //析构函数
-    QJsonObject toJson();               //把属性转化成Json格式
-    void setFromJson();                 //从Json格式改变属性值
-    void incRoomTmp();                  //室温升一度
-    void decRoomTmp();                  //室温降一度
+    Attribute();                            //构造函数
+    ~Attribute();                           //析构函数
+    QJsonObject toJson();                   //把属性转化成Json格式
+    void setFromJson(QByteArray byteArray); //从Json格式改变属性值
+    void incRoomTmp();                      //室温升一度
+    void decRoomTmp();                      //室温降一度
 
     //各属性的get和set方法
-    QString getRoomNum() const;
-    void setRoomNum(const QString &roomNum);
+    int getRoomNum() const;
+    void setRoomNum(const int &roomNum);
     int getWindSpeed() const;
     void setWindSpeed(const int &windSpeed);
-    float getRoomTmp() const;
+    double getRoomTmp() const;
     int getMode() const;
-    float getTargetTmp() const;
-    void setTargetTmp(float targetTmp);
+    void setMode(int mode);
+    double getTargetTmp() const;
+    void setTargetTmp(double targetTmp);
     bool getPower() const;
     void setPower(bool power);
+    void setIsServed(const bool isServed);
     bool getIsServed() const;
-    float getKWh() const;
-    //float getFee() const;
+
+    double getKWh() const;
+    double getFee() const;
 
 private:
-    QString _roomNum;                   //房间号
+    int _roomNum;                       //房间号
     int _windSpeed;                     //风速，低档，中档，高档
-    float _roomTmp;                     //室温，构造函数中=_defRoomTmp
+    double _roomTmp;                    //室温，构造函数中=_defRoomTmp
     int _mode;                          //工作模式，制冷，制热，启动之后从主机获取
-    float _targetTmp;                   //目标温度，构造函数中赋初值，启动之后从主机获取
+    double _targetTmp;                  //目标温度，构造函数中赋初值，启动之后从主机获取
     bool _power;                        //开关机
     bool _isServed;                     //是否正在被服务
-    float _defRoomTmp;                  //缺省室温，室温稳定值，构造函数中赋初值
-    float _kWh;                         //消耗的电量
-    //flaot _fee;                         //所需支付的费用
+    double _Kwh;                        //消耗的电量
+    double _fee;                        //所需支付的费用
 };
 
 #endif // ATTRIBUTE_H
