+++ clientblock.h
@@ -10,7 +10,7 @@ class ClientBlock : public QObject
 public slots:
     void readMessage();             //判断请求是否合理，包括目标温度当前温度是否合理，若接收到关机请求，则发送shutdown信号
 signals:
-    void shutdown(ClientBlock* );        //当接收到从控机的关机请求时，此信号用于通知Sever销毁自己
+    void shutdown(ClientBlock* );        //当接收到从控机的关机请求时，此信号用于通知Server销毁自己
     void update(QString roomNum);        //当收到消息时，触发此信号，主机面板上更新从机状态
 public:
     explicit ClientBlock(QTcpSocket* socket, int mode, QObject *parent = 0);//根据socket和主机工作模式构造CB
@@ -18,9 +18,9 @@ public:
     void check();                   //先判断是否服务完成，再判断是否被挂起（挂起不变），若不为0count--, 若count==0，变温度，再判断是否达到目标温度，改变isFinished,发消息，重制count，
     void sendMessage();             //向从控机发送消息，包括变温，和被挂起
     bool isSatisfied();             //是否服务完成
-    void setMode();                 //设置从机工作模式
+    void setMode(int mode);         //设置从机工作模式
     void setDefRoomTmp();           //设置从机缺省室温
-
+    void sendFirstMessage();
 private:
     void updateCount();             //count变为0时，或从机发送风速变化请求，需根据风速重新计算count
 private:
